+++ backend/src/main/java/com/example/easy/domain/guest/BillRepository.java
@@ -5,12 +5,12 @@ import com.example.easy.domain.Bill;
 import java.util.List;
 
 public interface BillRepository {
-    Bill loadById(int id);
+    Bill loadById(long id);
     void add(Bill bill);
     List<Bill> loadAll();
-    void delete(int id);
-    void update(int id,Bill bill);
-    Integer getIdByTableNumber(Integer table);
+    void delete(long id);
+    void update(long id,Bill bill);
+    Long getIdByTableNumber(Integer table);
 
     void deleteAll();
 }
